anticheat-agent-python/requirements.txt
requests
pynput
pystray
pillow
